airflow-docker/requirements.txt
faker
mysql-connector-python
pandas